third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_mark_with_name_of_navigation_timing_optional_attribute-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.mark("secureConnectionStart") threw an exception when secureConnectionStart attribute of Navigation Timing is supported. 
PASS window.performance.mark("secureConnectionStart") threw a SYNTAX_ERR when secureConnectionStart attribute of Navigation Timing is supported. 
Harness: the test ran to completion.
